CU-85zu0y2yu - Implement the backend filtering Refactored the registration endpoint Refactored the filtering logic
