+++ backend/src/main/java/com/example/backend/controller/IssueController.java
@@ -1,6 +1,6 @@
 package com.example.backend.controller;
 
-import com.example.backend.dto.filter.IssueFilter;
+import com.example.backend.dto.filter.FilterCriteria;
 import com.example.backend.dto.request.IssueCommentRequest;
 import com.example.backend.dto.request.IssueRequest;
 import com.example.backend.dto.response.IssueFullResponse;
@@ -48,7 +48,7 @@ public class IssueController {
     }
 
     @PostMapping(path = "/filter")
-    public ResponseEntity<List<IssueFullResponse>> getFilteredIssues(@RequestBody IssueFilter filter) {
+    public ResponseEntity<List<IssueFullResponse>> getFilteredIssues(@RequestBody FilterCriteria filter) {
         return new ResponseEntity<>(
                 mapper.toResponses(issueService.filterIssues(filter)),
                 OK
