+++ backend/src/main/java/com/example/backend/dao/IssueDao.java
@@ -1,6 +1,6 @@
 package com.example.backend.dao;
 
-import com.example.backend.dto.filter.IssueFilter;
+import com.example.backend.dto.filter.FilterCriteria;
 import com.example.backend.entity.issue.IssueEntity;
 
 import java.util.Optional;
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface IssueDao {
     List<IssueEntity> selectAllIssues();
-    List<IssueEntity> selectAllFilteredIssues(IssueFilter filter);
+    List<IssueEntity> selectAllFilteredIssues(FilterCriteria filter);
     Optional<IssueEntity> selectIssueByIssueId(String issueId);
     void insertIssue(IssueEntity issue);
     boolean existsIssueWithIssueId(String issueId);
