+++ backend/src/main/java/com/example/backend/service/IssueService.java
@@ -4,7 +4,7 @@ import com.example.backend.dao.IssueCommentDao;
 import com.example.backend.dao.IssueDao;
 import com.example.backend.dao.ProjectDao;
 import com.example.backend.dao.UserDao;
-import com.example.backend.dto.filter.IssueFilter;
+import com.example.backend.dto.filter.FilterCriteria;
 import com.example.backend.entity.ProjectEntity;
 import com.example.backend.entity.UserEntity;
 import com.example.backend.entity.issue.IssueCommentEntity;
@@ -54,7 +54,7 @@ public class IssueService {
         return issues;
     }
 
-    public List<IssueEntity> filterIssues(IssueFilter filter) {
+    public List<IssueEntity> filterIssues(FilterCriteria filter) {
 
         return issueDao.selectAllFilteredIssues(filter);
     }
